+++ src/main/java/graph/gedcom/MiniCard.java
@@ -12,12 +12,12 @@ public class MiniCard extends Card {
 	}
 
 	@Override
-	int centerRelX() {
+	float centerRelX() {
 		return width / 2;
 	}
 
 	@Override
-	public int centerX() {
+	public float centerX() {
 		return x + width / 2;
 	}
 }
